Added effects to basic turret
Basic turret now is affected by flowers as intended. Due to the projects structure I will have to go by hand to each turret and define how flower upgrades affect each one but that's alright for the scope of this project as all turrets are implemented already

+++ Tower Defense/Assets/Code/Scripts/Health.cs
@@ -12,7 +12,7 @@ public class Health : MonoBehaviour
 
 
     [Header("Attributes")]
-    [SerializeField] private int hitPoints = 2;
+    [SerializeField] private float hitPoints = 2f;
     [SerializeField] private int currencyWorth = 50;
     
     private GameObject childObj;
@@ -21,7 +21,7 @@ public class Health : MonoBehaviour
 
     private bool isDestroyed = false;
 
-    public void TakeDamage(int dmg)
+    public void TakeDamage(float dmg)
     {
         hitPoints -= dmg;
